backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
python-jose==3.3.0
pywebpush==2.0.0
cryptography==42.0.5
pydantic==2.7.1
